requirements.txt
absl-py==2.0.0
accelerate==1.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
audioread==3.0.1
av==13.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
blis==0.7.11
boto3==1.28.78
botocore==1.31.78
braceexpand==0.1.7
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
clean-fid==0.1.35
cleo==2.1.0
click==8.1.6
clip==1.0
cloudpathlib==0.15.1
cloudpickle==2.2.1
cmake==3.27.6
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
cos-python-sdk-v5==1.9.26
coscmd==1.8.6.33
cpm-kernels==1.0.11
crashtest==0.4.1
crcmod==1.7
cryptography==44.0.0
cubinlinker==0.3.0+2.gce0680b
cuda-python==12.2.0rc5+5.g84845d1
cudf==23.8.0
cugraph==23.8.0
cugraph-dgl==23.8.0
cugraph-service-client==23.8.0
cugraph-service-server==23.8.0
cuml==23.8.0
cupy-cuda12x==12.1.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.3
dask==2023.7.1
dask-cuda==23.8.0
dask-cudf==23.8.0
datasets==2.14.6
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.12.2
defusedxml==0.7.1
diffusers==0.31.0
dill==0.3.7
distlib==0.3.9
distributed==2023.7.1
dm-tree==0.1.8
dulwich==0.21.7
einops==0.7.0
en-core-web-sm==3.7.0
exceptiongroup==1.1.3
execnet==2.0.2
executing==2.0.0
expecttest==0.1.3
fairscale==0.4.13
fastChat==0.1.1
fastjsonschema==2.18.1
fastrlock==0.8.1
filelock==3.12.4
flash-attn==2.6.1
flatbuffers==24.3.25
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==0.4.6
graphsurgeon==0.4.6
grpcio==1.59.0
hjson==3.1.0
huggingface-hub==0.23.4
humanfriendly==10.0
hypothesis==5.35.1
idna==3.4
image-reward==1.5
imageio==2.34.2
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
intel-openmp==2021.4.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter-tensorboard==0.2.0
jupyterlab==2.3.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupytext==1.15.2
keyring==24.3.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.40.1
locket==1.0.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==5.1.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
ninja==1.11.1.1
notebook==6.4.10
numba==0.57.1+1.g5fba9aa8f
numpy==1.26.4
nvfuser==0.0.20+gitunknown
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda120==1.30.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
nvidia-pyindex==1.0.9
nvtx==0.2.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.0
onnx-graphsurgeon==0.3.27
onnxruntime-gpu==1.15.1
open-clip-torch==2.23.0
opencv==4.7.0
opencv-contrib-python==4.10.0.84
opencv-python==4.8.0.74
optimum==1.19.2
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathy==0.10.2
peft==0.13.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pillow-heif==0.13.1
pip==23.2.1
pkginfo==1.11.2
platformdirs==3.11.0
pluggy==1.3.0
ply==3.11
poetry-core==1.9.1
poetry-plugin-export==1.8.0
polygraphy==0.49.0
pooch==1.7.0
portalocker==2.10.0
preshed==3.0.9
prettytable==3.9.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.4
ptxcompiler==0.8.1+1.g2cb1b35
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pybind11-global==2.11.1
pycocotools==2.0+nv0.7.3
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pylibcugraph==23.8.0
pylibcugraphops==23.8.0
pylibraft==23.8.0
pynvml==11.4.1
pyparsing==3.1.1
pyproject_hooks==1.2.0
pytest==7.4.2
pytest-flakefinder==1.1.0
pytest-rerunfailures==12.0
pytest-shard==0.1.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytorch-quantization==2.1.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
qwen-vl-utils==0.0.8
raft-dask==23.8.0
RapidFuzz==3.10.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
resampy==0.4.2
rich==13.9.4
rmm==23.8.0
rpds-py==0.10.4
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.2.0
scipy==1.11.1
seaborn==0.13.0
SecretStorage==3.3.3
Send2Trash==1.8.2
sentencepiece==0.1.99
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx-glpi-theme==0.3
srsly==2.4.8
stable-fast==0.0.13.post4
stack-data==0.6.3
SwissArmyTransformer==0.4.8
sympy==1.12
tabulate==0.9.0
tbb==2021.10.0
tblib==2.0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorrt==8.6.1
terminado==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
thriftpy2==0.4.16
tifffile==2024.7.2
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
toolz==0.12.0
torch==2.3.0+cu121
torch-tensorrt==0.0.0
torchdata==0.7.0a0
torchtext==0.16.0a0
torchvision==0.18.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.46.3
treelite==3.2.0
treelite-runtime==3.2.0
triton==2.3.0
trove-classifiers==2024.10.21.16
typer==0.9.0
types-dataclasses==0.6.6
typing_extensions==4.9.0
ucx-py==0.33.0
uff==0.6.9
urllib3==1.26.16
virtualenv==20.28.0
voluptuous==0.14.2
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.2
webdataset==0.2.73
webencodings==0.5.1
Werkzeug==3.0.0
wheel==0.41.2
xdoctest==1.0.2
xformers==0.0.26.post1
xgboost==1.7.5
xhsBaseCV==0.0.1
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.2
zict==3.0.0
zipp==3.16.2
zope.interface==6.1
